Fix 64-bit time_t on 32-bit architectures

+++ src/db_file.c
@@ -1,7 +1,7 @@
 /*
  * AIDE (Advanced Intrusion Detection Environment)
  *
- * Copyright (C) 1999-2007, 2010-2013, 2016, 2018-2023 Rami Lehti,
+ * Copyright (C) 1999-2007, 2010-2013, 2016, 2018-2024 Rami Lehti,
  *               Pablo Virolainen, Mike Markley, Richard van den Berg,
  *               Hannes von Haugwitz
  *
@@ -445,7 +445,7 @@ int db_write_time_base64(time_t i,FILE* file,int a)
 
   memset((void*)ptr,0,len);
 
-  snprintf(ptr, len, "%li",i);
+  snprintf(ptr, len, "%llu", (unsigned long long) i);
 
 
   tmpstr=encode_base64((byte *)ptr,strlen(ptr));
